set stream server
