+++ cloud/pkg/cloudstream/session.go
@@ -1,4 +1,4 @@
-package streamcontroller
+package cloudstream
 
 import (
 	"fmt"
